backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.0.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
google-search-results==2.4.2
elevenlabs==0.2.26

